Remove the outdate no_filter_ui section
The `no_filter_ui` section on the `index_page.html` was never used.
The only reference to it was in the `BlogIndexPage` and only to prevent
the section from rendering.

Since we don't need it anywhere, it feels cleaner to remove this section
entirely.

This is the references to `no_filter_ui` that I had found:
```console
$ ag no_filter_ui
network-api/networkapi/wagtailpages/templates/wagtailpages/index_page.html
32:    {% if filtered and no_filter_ui != True %}  {# Start of IF for filtered page content. Note that this injects an entire row. #}

network-api/networkapi/wagtailpages/pagemodels/blog/blog_index.py
113:        context['no_filter_ui'] = True
```